% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{cal_append_current}
\alias{cal_append_current}
\title{Add the current calendar to the upcoming calendar.}
\usage{
cal_append_current(data, calendar_url = epi_calendar_url)
}
\arguments{
\item{data}{Data with columns that match standard iCal format.}

\item{calendar_url}{The URL to the `.ics` file.}
}
\value{
A [tibble::tibble()].
}
\description{
Add the current calendar to the upcoming calendar.
}
