% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{cal_create_event}
\alias{cal_create_event}
\title{Create a tibble with an iCal event.}
\usage{
cal_create_event(start, end, title, description)
}
\arguments{
\item{start}{The datetime of when the event starts. Must be `POSIXct` or similar, recommended to use [lubridate::as_datetime()] to correctly format the date.}

\item{end}{The datetime of when the event ends. See `start` for formatting recommendations.}

\item{title}{The title shown in the calendar event.}

\item{description}{A description that is shown in the calendar event.}
}
\value{
A single row of a [tibble::tibble()].
}
\description{
Create a tibble with an iCal event.
}
\examples{
cal_create_event(
  start = lubridate::as_datetime(lubridate::now()),
  end = lubridate::now() + lubridate::hours(1),
  title = "Meeting",
  description = "Discussing the new project"
)
}
