% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{cal_read_ical}
\alias{cal_read_ical}
\title{Read in a iCal file or URL}
\usage{
cal_read_ical(ics)
}
\arguments{
\item{ics}{A file or URL that is in the `.ics` format.}
}
\value{
A [tibble::tibble()].
}
\description{
Read in a iCal file or URL
}
