% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{cal_set_epi_meeting_details}
\alias{cal_set_epi_meeting_details}
\title{Add calendar details for Epi Group meetings and standardize to the iCal format.}
\usage{
cal_set_epi_meeting_details(data)
}
\arguments{
\item{data}{Data with columns that match standard iCal format.}
}
\value{
A [tibble::tibble()].
}
\description{
Add calendar details for Epi Group meetings and standardize to the iCal format.
}
