% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{cal_write_ical}
\alias{cal_write_ical}
\title{Save data as a file in the iCal format.}
\usage{
cal_write_ical(data)
}
\arguments{
\item{data}{Data with columns that match standard iCal format.}
}
\value{
A file.
}
\description{
File is saved to `inst/calendar.ics`.
}
