% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{epi_add_sessions}
\alias{epi_add_sessions}
\title{Add new events to the current calendar.}
\usage{
epi_add_sessions(current, new)
}
\arguments{
\item{current}{The current calendar.}

\item{new}{New events to add to the calendar.}
}
\value{
A [tibble::tibble()].
}
\description{
Right now is mostly a wrapper around `dplyr::full_join()`.
}
