% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{epi_create_session_issues}
\alias{epi_create_session_issues}
\title{Add upcoming sessions to the GitHub Project Board as GitHub Issues.}
\usage{
epi_create_session_issues(dates)
}
\arguments{
\item{dates}{The dates for the upcoming sessions.}
}
\value{
A GitHub JSON response object.
}
\description{
Add upcoming sessions to the GitHub Project Board as GitHub Issues.
}
