% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{epi_create_sessions}
\alias{epi_create_sessions}
\title{Create default sessions dates for Epi Group meetings.}
\usage{
epi_create_sessions(start, end)
}
\arguments{
\item{start}{The datetime of when the event starts. Must be `POSIXct` or similar, recommended to use [lubridate::as_datetime()] to correctly format the date.}

\item{end}{The datetime of when the event ends. See `start` for formatting recommendations.}
}
\value{
A [tibble::tibble()].
}
\description{
These are standardize to the iCal format. It's important to use the
correct timezone for the dates when giving them to this function.
}
\examples{
upcoming <- lubridate::as_datetime(c(
  "2024-01-06 13:00:00",
  "2024-02-03 13:00:00"
), tz = "Europe/Copenhagen")

epi_create_events(
  upcoming,
  upcoming + lubridate::hours(1)
)

}
