% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{epi_get_calendar}
\alias{epi_get_calendar}
\title{Get's the Epi calendar and outputs as a tibble.}
\usage{
epi_get_calendar(path = epi_calendar_url)
}
\arguments{
\item{path}{The path or URL to the file that is in the `.ics` format.}
}
\value{
A [tibble::tibble()].
}
\description{
Get's the Epi calendar and outputs as a tibble.
}
\examples{
\dontrun{
epi_get_calendar()
}
}
