% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{gh_create_issue}
\alias{gh_create_issue}
\title{Create a new GitHub issue on the SDCA GitHub organization.}
\usage{
gh_create_issue(
  repo,
  title,
  body = NA,
  assignee = NA,
  template = NA,
  label = NA,
  project = NA
)
}
\arguments{
\item{repo}{The name of the repo in the SDCA GitHub organization.}

\item{title}{The title to give the new issue.}

\item{body}{The content of the issue itself.}

\item{assignee}{Who to assign the issue.}

\item{template}{Which template to use for the issue's body.}

\item{label}{The label to give the issue.}

\item{project}{The Project Board to add the issue to.}
}
\value{
A GitHub JSON response object.
}
\description{
Create a new GitHub issue on the SDCA GitHub organization.
}
