% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{read_ical}
\alias{read_ical}
\title{Read in an iCal file as a tibble.}
\usage{
read_ical(path)
}
\arguments{
\item{path}{The path to the `.ics` file.}
}
\value{
A [tibble::tibble()].
}
\description{
Read in an iCal file as a tibble.
}
