% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-trello-cards.R
\name{tr_auth_epi}
\alias{tr_auth_epi}
\title{Authenticate with the Epi Group Trello board.}
\usage{
tr_auth_epi()
}
\value{
A token used to connect and authenticate with Trello.
}
\description{
Details about getting Key and OAuth Secret are found on this
[website](https://developer.atlassian.com/cloud/trello/guides/rest-api/authorization/).
}
