% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-trello-cards.R
\name{tr_get_board_id}
\alias{tr_get_board_id}
\title{Get the ID for a Trello board.}
\usage{
tr_get_board_id(url = epi_board_url, board_name = "Upcoming", token)
}
\arguments{
\item{url}{The board URL.}

\item{board_name}{The name of the board you want the ID for.}

\item{token}{The token obtained from [tr_auth_epi()].}
}
\value{
A character vector of the board ID.
}
\description{
Get the ID for a Trello board.
}
