% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-trello-cards.R
\name{tr_read_card_template}
\alias{tr_read_card_template}
\title{Get the card template to use when making Trello cards.}
\usage{
tr_read_card_template()
}
\value{
A character vector of the card template.
}
\description{
Get the card template to use when making Trello cards.
}
