% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{update_epi_calendar_meetings}
\alias{update_epi_calendar_meetings}
\title{Update the Epi Group calendar with upcoming meetings.}
\usage{
update_epi_calendar_meetings(data)
}
\arguments{
\item{data}{Upcoming meetings for the Epi Group.}
}
\value{
A `.ics` file of upcoming meetings.
}
\description{
Update the Epi Group calendar with upcoming meetings.
}
