% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-trello-cards.R
\name{update_epi_trello_meetings}
\alias{update_epi_trello_meetings}
\title{Add upcoming meetings to Trello as cards.}
\usage{
update_epi_trello_meetings(data, token)
}
\arguments{
\item{data}{Data that contains the month and day of the upcoming meetings.}

\item{token}{The token obtained from [tr_auth_epi()].}
}
\value{
Used for the side effect of posting to Trello.
}
\description{
Add upcoming meetings to Trello as cards.
}
