% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{write_ical}
\alias{write_ical}
\title{Save data as a file in the iCal format.}
\usage{
write_ical(data, path = NA)
}
\arguments{
\item{data}{The [tibble::tibble()] that contains the events to save in iCal format.}

\item{path}{The path to the output file.}
}
\value{
A file.
}
\description{
File is saved to `inst/epi-calendar.ics`.
}
